remove portal_url: fix skin directory related issues
getToolByName return a real utility and this doesn't works with
the security stuff so it raise unauthorized on any call to the
utility.

It seems to works when I'm using the browser view their, so
I have changed every calls to use browser view.

An other try has been done making the utility inherits from
a SimpleItem but in that case the security breaks when you try
to use it in zope3 context (like in a vocabulary for example).

We almost don't care of skins related issues because they should
be removed in an other plips.