remove portal_url: fix testURLTool
the tool is not persistent anymore so we don't care of the acquisition.
But the components are not loaded here so the fake plone site is not
a component registry.

I ran out of time here, so I just don't tests the components call
to get the site, so I fill the cache with the dummy site.